finish

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: yujelee <yujelee@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/06 12:55:21 by yujelee           #+#    #+#             */
-/*   Updated: 2022/07/08 20:46:28 by yujelee          ###   ########.fr       */
+/*   Updated: 2022/07/11 19:31:19 by yujelee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (str[count])
